+++ eventqueue/eventqueue-common/src/main/kotlin/Event.kt
@@ -1,0 +1,4 @@
+package com.redmadrobot.eventqueue
+
+/** Marker interface for entities that can be put to the [EventQueue]. */
+public interface Event